Rename: 모델 클래스 model 디렉터리로 이동
