+++ src/main/java/ojosama/talkak/comment/CommentRepository.java
@@ -1,6 +1,7 @@
 package ojosama.talkak.comment;
 
 import java.util.List;
+import ojosama.talkak.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
