+++ src/main/java/ojosama/talkak/model/Comment.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.comment;
+package ojosama.talkak.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,8 +8,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import ojosama.talkak.user.Member;
-import ojosama.talkak.video.Video;
 
 @Entity
 @Table(name = "comment")
