+++ src/main/java/ojosama/talkak/model/Member.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.user;
+package ojosama.talkak.model;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
@@ -9,7 +9,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.util.List;
-import ojosama.talkak.comment.Comment;
 
 @Entity
 @Table(name = "member")
