+++ src/main/java/ojosama/talkak/model/Reaction.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.reaction;
+package ojosama.talkak.model;
 
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
@@ -6,8 +6,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import ojosama.talkak.user.Member;
-import ojosama.talkak.video.Video;
+import ojosama.talkak.reaction.ReactionId;
 
 @Entity
 @Table(name = "reaction")
