+++ src/main/java/ojosama/talkak/reaction/ReactionRepository.java
@@ -1,5 +1,6 @@
 package ojosama.talkak.reaction;
 
+import ojosama.talkak.model.Reaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
