+++ src/main/java/ojosama/talkak/user/MemberRepository.java
@@ -1,5 +1,6 @@
 package ojosama.talkak.user;
 
+import ojosama.talkak.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
