+++ src/main/java/ojosama/talkak/video/VideoRepository.java
@@ -1,5 +1,6 @@
 package ojosama.talkak.video;
 
+import ojosama.talkak.model.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
